requirements.txt
# Google Gemini API (para información e imágenes)
google-genai==0.5.0

# Web Framework
fastapi==0.108.0
uvicorn==0.25.0
jinja2==3.1.2
python-multipart==0.0.6
mangum==0.17.0

# HTTP Client for external API calls
httpx==0.26.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Session storage
redis==5.0.1

# Image processing (for web display)
pillow>=10.0.0
